Push `trans_unit` to `ctx` earlier to make sure `trans_unit` is live until the end
The old version can only guarantee that `trans_unit` is live until `getcursor` has returned. The `root_cursor` is a kind of `CLCursor` which is an immutable struct that only holds a underlying `CXCursor` pointer. The resource that being pointed could be invalid if the `trans_unit` were dead too early.