Only check locations when dealing with anonymous tags
fix #405

This bug is quite subtle. The current implementation inserts unique ids for anonymous tags(Julia doesn't support anonymous structs anyway, so we must do this), these anonymous tags could be in different translation units, and they could also be defined in two different source files(with exactly the same tokens). In the latter case, as we don't know which typedef id is kept, we are not able to determine which dependent anonymous tag should be kept, except the situation where two anonymous tags share the same source location.