Avoid method ambiguities in TypedefMutualRef's
`emit!()` for `TypedefMutualRef` will create some specializations of
`Base.unsafe_convert()`, and previously they didn't declare a type for the
`Ref`/`Ptr` value being passed to them, which would cause a method ambiguity
with Base (detected by Aqua.jl).

The problem is that the typedef will of course be declared in the C header
before the struct it references, so the `Base.unsafe_convert()` specializations
couldn't use the actual struct type in their signature because it won't have
been created yet. This commit fixes that with the notion of a *partially emitted
node* that contains *premature exprs* that haven't been emitted yet. Now
`emit!()` for `TypedefMutualRef` will record the `Base.unsafe_convert()` methods
as premature exprs and mark itself as a partially emitted node. Then later,
`emit!()` for the `StructDefinition` will look for any partially emitted nodes
that are marked for it and emit their premature expressions itself.