Disable libclang's CrashRecoveryContext
LLVM's CrashRecoveryContext replaces all signal handlers and assumes that
a SIGSEGV signal is fatal, which is incompatible with language runtimes
like Julia / Java which use this signal as a hook for GC safepoints.

libclang does not provide any mechanism to disable this behavior (which
affects the application globally), except via the "LIBCLANG_DISABLE_CRASH_RECOVERY"
environment variable. Sadly changing ENV on-the-fly is not thread-safe
either, but it's the best we can do to avoid losing our signal handlers.